order match logic, trade class and dto, order factory

+++ src/main/resources/db/migration/V1__create_order_table.sql
@@ -7,6 +7,6 @@ CREATE TABLE "orders"
   operation_type VARCHAR(5) NOT NULL,
   ticker_name VARCHAR(5) NOT NULL,
   quantity INT NOT NULL,
-  price DECIMAL NOT NULL,
+  price BIGINT NOT NULL,
   created_at TIMESTAMP NOT NULL
 );
